TeraType and GmaxFactor conditions
Added TeraTypeCondition - checks if the Pokemon's Tera Type is in the list of provided types. Can be inverted with field 'invert'. (defaults to false)

Added GmaxFactorCondition - checks if the Pokemon has the provided Gigantamax Factor state. true for has, false for does not.

Type table examples for TeraType and GmaxFactor conditions.

Signed-off-by: anabsolutesloth <63694982+anabsolutesloth@users.noreply.github.com>